refactor: 함수들을 Application파일에서 RacingCar 클래스로 이동
